build/windows: Do not bundle unneeded share/metainfo on Windows
Seems it was used at runtime in the past but that is
not the case anymore. Indeed, it started to be bundled in
af79bbe028a41a16ce66574447f011dc4b583079 due to an
error: before the Windows bundling unification the
crossbuilds "script" on .yml bundled all the 'share' dir.

(cherry picked from commit 472123181de5aba805614726b00d6b3cdc02b5d9)