app: fix #13555 crash in clipboard when closing
Apparently the gimpclipboard_send_* functions can be called when
gimp_clip has already been cleared, usually at shutdown. This is
causing crashes sometimes that I have not been able to reproduce.

Fixing this at the source would be best, but I have not been able to
figure out where that would be. So, next best thing is adding a check
to bail out early when gimp_clip is NULL.